+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # v0.2.0 (unreleased)
-- new: easily create knobs from image strips using `<ImageKnob />`
+- new: easily create knobs from image strips using `<ImageStripControl />`
 - new: helper function for creating smoothly animated movement: `createSmoothedValue(value: () => number, speed = 1)`
 - fix: the value jumped when pressing shift while already dragging a control